Работа по замечаниям
