+++ ImageFeed/Auth/WebViewViewController.swift
@@ -8,22 +8,13 @@
 import UIKit
 @preconcurrency import WebKit
 
-private enum WebViewConstants {
-    static let unsplashAuthorizeURLString = "https://unsplash.com/oauth/authorize"
-}
-
-protocol WebViewViewControllerDelegate: AnyObject {
-    func webViewViewController(_ vc: WebViewViewController, didAuthenticateWithCode code: String)
-    func webViewViewControllerDidCancel(_ vc: WebViewViewController)
-}
-
 
 final class WebViewViewController: UIViewController {
     
     weak var delegate: WebViewViewControllerDelegate?
-    private var webView: WKWebView!
-    private var backButton: UIButton!
-    private var progressView: UIProgressView!
+    private var webView: WKWebView = WKWebView()
+    private var backButton: UIButton = UIButton()
+    private var progressView: UIProgressView = UIProgressView()
     private var estimatedProgressObservation: NSKeyValueObservation?
     
     
@@ -37,45 +28,43 @@ final class WebViewViewController: UIViewController {
         
         estimatedProgressObservation = webView.observe(
             \.estimatedProgress,
-            options: [],
-            changeHandler: { [weak self] _, _ in
-                guard let self = self else { return }
-                self.updateProgress()
-            }
+             options: [],
+             changeHandler: { [weak self] _, _ in
+                 guard let self = self else { return }
+                 self.updateProgress()
+             }
         )
     }
-
+    
     private func setupWebView() {
-        webView = WKWebView(frame: view.bounds)
         webView.translatesAutoresizingMaskIntoConstraints = false
         view.addSubview(webView)
     }
     
     private func setupProgressView() {
-        progressView = UIProgressView()
         progressView.progress = 0.5
         progressView.tintColor = UIColor(red: 26, green: 27, blue: 34, alpha: 1)
         progressView.translatesAutoresizingMaskIntoConstraints = false
         view.addSubview(progressView)
     }
-
+    
     private func loadAuthView() {
-        guard var urlComponents = URLComponents(string: WebViewConstants.unsplashAuthorizeURLString) else {
+        guard var urlComponents = URLComponents(string: Constants.unsplashAuthorizeURLString) else {
             print("Строка подключения не связана с URLComponents")
             return
         }
-
+        
         urlComponents.queryItems = [
             URLQueryItem(name: "client_id", value: Constants.accessKey),
             URLQueryItem(name: "redirect_uri", value: Constants.redirectURI),
             URLQueryItem(name: "response_type", value: "code"),
             URLQueryItem(name: "scope", value: Constants.accessScope)
         ]
-
+        
         guard let url = urlComponents.url else {
             return
         }
-
+        
         let request = URLRequest(url: url)
         webView.load(request)
         
@@ -100,7 +89,7 @@ final class WebViewViewController: UIViewController {
         super.viewDidDisappear(animated)
         delegate?.webViewViewControllerDidCancel(self)
     }
-
+    
     private func updateProgress() {
         progressView.progress = Float(webView.estimatedProgress)
         progressView.isHidden = fabs(webView.estimatedProgress - 1.0) <= 0.0001
@@ -110,33 +99,4 @@ final class WebViewViewController: UIViewController {
         estimatedProgressObservation = nil
     }
     
-}
-
-extension WebViewViewController: WKNavigationDelegate {
-    func webView(
-        _ webView: WKWebView,
-        decidePolicyFor navigationAction: WKNavigationAction,
-        decisionHandler: @escaping (WKNavigationActionPolicy) -> Void
-    ) {
-         if let code = code(from: navigationAction) {
-                UIBlockingProgressHUD.show()
-                delegate?.webViewViewController(self, didAuthenticateWithCode: code)
-                decisionHandler(.cancel)
-          } else {
-                decisionHandler(.allow)
-            }
-    }
-    private func code(from navigationAction: WKNavigationAction) -> String? {
-        if
-            let url = navigationAction.request.url,
-            let urlComponents = URLComponents(string: url.absoluteString),
-            urlComponents.path == "/oauth/authorize/native",
-            let items = urlComponents.queryItems,
-            let codeItem = items.first(where: { $0.name == "code" })
-        {
-            return codeItem.value
-        } else {
-            return nil
-        }
-    }
 }
