+++ ImageFeed/Constants.swift
@@ -18,4 +18,5 @@ enum Constants {
         }
         return url
     }()
+    static let unsplashAuthorizeURLString = "https://unsplash.com/oauth/authorize"
 }
