+++ ImageFeed/ImagesList/ImagesListViewController.swift
@@ -11,7 +11,7 @@ final class ImagesListViewController: UIViewController {
     
     @IBOutlet private var tableView: UITableView!
     private let showSingleImageSegueIdentifier = "ShowSingleImage"
-
+    
     private let photosName: [String] = Array(0..<20).map{ "\($0)" }
     private lazy var dateFormatter: DateFormatter = {
         let formatter = DateFormatter()
@@ -41,7 +41,7 @@ final class ImagesListViewController: UIViewController {
             assertionFailure("Invalid segue destination")
             return
         }
-
+        
         let image = UIImage(named: photosName[indexPath.row])
         viewController.image = image
     }
