+++ ImageFeed/Services/HTTPMethod.swift
@@ -1,0 +1,13 @@
+//
+//  HTTPMethod.swift
+//  ImageFeed
+//
+//  Created by Денис Кель on 07.03.2025.
+//
+
+import Foundation
+
+enum HTTPMethod: String {
+    case get = "GET"
+    case post = "POST"
+}
