+++ ImageFeed/Services/Profile.swift
@@ -1,0 +1,22 @@
+//
+//  Profile.swift
+//  ImageFeed
+//
+//  Created by Денис Кель on 07.03.2025.
+//
+
+import Foundation
+
+struct Profile {
+    let username: String
+    let name: String
+    let loginName: String
+    let bio: String?
+    
+    init(profileResult: ProfileResult) {
+        self.username = profileResult.username
+        self.name = "\(profileResult.firstName ?? "") \(profileResult.lastName ?? "")".trimmingCharacters(in: .whitespaces)
+        self.loginName = "@\(profileResult.username)"
+        self.bio = profileResult.bio
+    }
+}
