+++ ImageFeed/Services/ProfileImageService.swift
@@ -5,76 +5,76 @@
 //  Created by Денис Кель on 05.03.2025.
 //
 
-import UIKit
 import Foundation
 
 final class ProfileImageService {
     static let shared = ProfileImageService()
     static let didChangeNotification = Notification.Name(rawValue: "ProfileImageProviderDidChange")
-
     private var avatarURL: String?
     private var task: URLSessionTask?
-    private init() {}
-
+    
+    
     private struct UserResult: Codable {
         let profileImage: ProfileImage
-
+        
         enum CodingKeys: String, CodingKey {
             case profileImage = "profile_image"
         }
     }
-
+    
     private struct ProfileImage: Codable {
         let small: String
     }
-
+    
+    private init() {}
+    
     func fetchProfileImageURL(username: String, _ completion: @escaping (Result<String, Error>) -> Void) {
-            task?.cancel()
-            guard let token = OAuth2TokenStorage.shared.token else {
-                print("[ProfileImageService]: Unauthorized request - missing token")
-                completion(.failure(NetworkError.unauthorized))
-                return
-            }
-            guard let request = makeProfileImageRequest(username: username, token: token) else {
-                print("[ProfileImageService]: Failed to create request for username \(username)")
-                completion(.failure(NetworkError.invalidRequest))
-                return
-            }
-
-            let task = URLSession.shared.objectTask(for: request) { [weak self] (result: Result<UserResult, Error>) in
-                DispatchQueue.main.async {
-                    guard let self = self else { return }
-                    self.task = nil
-
-                    switch result {
-                    case .success(let userResult):
-                        let avatarURL = userResult.profileImage.small
-                        self.avatarURL = avatarURL
-                        completion(.success(avatarURL))
-                        NotificationCenter.default.post(
-                            name: ProfileImageService.didChangeNotification,
-                            object: self,
-                            userInfo: ["URL": avatarURL]
-                        )
-                    case .failure(let error):
-                        print("[ProfileImageService]: Profile image fetch failed - \(error.localizedDescription)")
-                        completion(.failure(error))
-                    }
+        task?.cancel()
+        guard let token = OAuth2TokenStorage.shared.token else {
+            print("[ProfileImageService]: Unauthorized request - missing token")
+            completion(.failure(NetworkError.unauthorized))
+            return
+        }
+        guard let request = makeProfileImageRequest(username: username, token: token) else {
+            print("[ProfileImageService]: Failed to create request for username \(username)")
+            completion(.failure(NetworkError.invalidRequest))
+            return
+        }
+        
+        let task = URLSession.shared.objectTask(for: request) { [weak self] (result: Result<UserResult, Error>) in
+            DispatchQueue.main.async {
+                guard let self = self else { return }
+                self.task = nil
+                
+                switch result {
+                case .success(let userResult):
+                    let avatarURL = userResult.profileImage.small
+                    self.avatarURL = avatarURL
+                    completion(.success(avatarURL))
+                    NotificationCenter.default.post(
+                        name: ProfileImageService.didChangeNotification,
+                        object: self,
+                        userInfo: ["URL": avatarURL]
+                    )
+                case .failure(let error):
+                    print("[ProfileImageService]: Profile image fetch failed - \(error.localizedDescription)")
+                    completion(.failure(error))
                 }
             }
-
-            self.task = task
-            task.resume()
+        }
+        
+        self.task = task
+        task.resume()
     }
-
+    
     private func makeProfileImageRequest(username: String, token: String) -> URLRequest? {
         guard let url = URL(string: "https://api.unsplash.com/users/\(username)") else {
             print("Ошибка: Не удалось создать baseURL")
             return nil
         }
-
+        
         var request = URLRequest(url: url)
-        request.httpMethod = "GET"
+        request.httpMethod = HTTPMethod.get.rawValue
         request.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
         return request
     }
