+++ ImageFeed/SingleImage/SingleImageViewController.swift
@@ -15,11 +15,11 @@ final class SingleImageViewController: UIViewController {
             imageView.image = image
         }
     }
-
+    
     // IBOutlets
     @IBOutlet private weak var imageView: UIImageView!
     @IBOutlet private weak var scrollView: UIScrollView!
-
+    
     // IBActions
     @IBAction private func didTapShareButton(_ sender: Any) {
         guard let image = image else {
@@ -33,11 +33,11 @@ final class SingleImageViewController: UIViewController {
         activityViewController.overrideUserInterfaceStyle = .dark
         present(activityViewController, animated: true)
     }
-
+    
     @IBAction private func didTapBackButton(_ sender: Any) {
         dismiss(animated: true)
     }
-
+    
     // Методы жизненного цикла
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -47,34 +47,34 @@ final class SingleImageViewController: UIViewController {
         updateImageScaleAndPosition()
         scrollView.delegate = self
     }
-
+    
     // Приватные методы
     private func updateImageScaleAndPosition() {
         guard let image = image else { return }
-
+        
         let minZoomScale = scrollView.minimumZoomScale
         let maxZoomScale = scrollView.maximumZoomScale
         let visibleRectSize = scrollView.bounds.size
         let imageSize = image.size
-
+        
         guard !visibleRectSize.width.isZero, !visibleRectSize.height.isZero,
               !imageSize.width.isZero, !imageSize.height.isZero else {
             return
         }
-
+        
         let hScale = visibleRectSize.width / imageSize.width
         let vScale = visibleRectSize.height / imageSize.height
         let scale = min(maxZoomScale, max(minZoomScale, min(hScale, vScale)))
-
+        
         scrollView.setZoomScale(scale, animated: false)
         scrollView.layoutIfNeeded()
-
+        
         let newContentSize = CGSize(
             width: imageSize.width * scale,
             height: imageSize.height * scale
         )
         scrollView.contentSize = newContentSize
-
+        
         let x = (newContentSize.width - visibleRectSize.width) / 2
         let y = (newContentSize.height - visibleRectSize.height) / 2
         scrollView.setContentOffset(CGPoint(x: x, y: y), animated: false)
